+++ src/ReadoutEquipment.h
@@ -114,6 +114,7 @@ class ReadoutEquipment
   std::vector<CounterStats> equipmentStats;
   std::vector<CounterValue> equipmentStatsLast;
   std::bitset<RdhMaxLinkId + 1> equipmentLinksUsed;
+  std::vector<uint64_t> equipmentLinksData;
 
   double cfgConsoleStatsUpdateTime = 0;     // number of seconds between regular printing of statistics on console (if zero, only on stop)
   AliceO2::Common::Timer consoleStatsTimer; // timer to keep track of elapsed time between console statistics updates